# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Dapp",
        "description": "Dapp application rest api",
        "contact": {
            "email": "babichev.oleg.n@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/words/search": {
            "post": {
                "tags": [
                    "Words"
                ],
                "summary": "Search words by value",
                "description": "Search words by value",
                "operationId": "searchWords",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "val": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "val": "test"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/api/wordprogress": {
            "get": {
                "tags": [
                    "Word progress"
                ],
                "summary": "Get word progresses list",
                "description": "Get word progresses list",
                "operationId": "getWordProgresses",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Word progress"
                ],
                "summary": "Create word progress",
                "description": "Create word progress",
                "operationId": "wordprogressPost",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "learned": {
                                        "type": "bool"
                                    },
                                    "word_id": {
                                        "type": "integer"
                                    },
                                    "translation_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "learned": true,
                                    "word_id": 362742,
                                    "translation_id": 1
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Resource not found"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/api/wordprogress?filter=repeat": {
            "get": {
                "tags": [
                    "Word progress"
                ],
                "summary": "Get word progresses to repeat",
                "description": "Get word progresses to repeat",
                "operationId": "getRepeatWordProgresses",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/api/wordprogress/{wordprogressId}?action=repeat": {
            "put": {
                "tags": [
                    "Word progress"
                ],
                "summary": "Repeat word progress",
                "description": "Repeat word progress",
                "operationId": "wordProgressRepeatPut",
                "parameters": [
                    {
                        "name": "wordprogressId",
                        "in": "path",
                        "description": "ID of word progress",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "success": {
                                        "type": "bool"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "success": true
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Resource not found"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "api_key": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}